Documentation and Files
Updated README to include new wrappers, also deleted old NaN and scale action wrapper files and moved wrappers into AEC wrappers.
