+++ README.md
@@ -55,6 +55,14 @@ You can install SuperSuit via `pip install supersuit`
 
 `resize_v0(env, x_size, y_size, linear_interp=False)` Performs interpolation to up-size or down-size observation image using area interpolation by default. Linear interpolation is also available by setting `linear_interp=True` (it's faster and better for up-sizing). This wrapper is only available for 2D or 3D observations, and only makes sense if the observation is an image.
 
+`nan_noop_wrapper(env)` If an action is a NaN value the following function will perform a no operation action in its place.
+
+`nan_zeros_wrapper(env)` If an action is a NaN value the following function will perform a zeros operation action in its place.
+
+`nan_random_wrapper(env)` If an action is a NaN value the following function will perform a random operation action in its place. The random action will be retrieved from the action mask.
+
+`scale_actions_wrapper(env)` Scales the high and low bounds of the action space by a given value. Additionally, scales any actions by the same value.
+
 
 ## Included Multi-Agent Only Functions
